% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_bayesian_mpra.R
\name{est_sample_params}
\alias{est_sample_params}
\title{Estimate Transfection Parameters}
\usage{
est_sample_params(count_dat)
}
\arguments{
\item{count_dat}{}
}
\description{
Estimate Transfection Parameters for a single allele
}
