% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_bayesian_mpra.R
\name{find_weights}
\alias{find_weights}
\title{Produce annotation-based weightings}
\usage{
find_weights(i, dist_mat, min_dist_kernel, min_num_contributing = 30,
  increase_fold = 1.333)
}
\arguments{
\item{i}{the index of the variant in the assay}

\item{dist_mat}{annotation-based distance matrix}

\item{min_dist_kernel}{the minimum kernel possible to use in the annotation
space}

\item{min_num_contributing}{the minimum number of variants that must
contribute to the ith variant's prior}

\item{increase_fold}{the multiplicative amount by which to increase the
kernel in the case the current kernel doesn't allow at least
\code{min_num_contributing} variants to contribute}
}
\value{
a similarity-based vector of weights of the other n-1 variants
}
\description{
For the ith variant, adaptively produce a vector of weightings such that at minimum
min_num_contributing variants meaningfully contribute (i.e.
cumsum(sortedWeights)[30] <= .99) arguments after the 2nd are heuristics that
may need tuning
}
\details{
The kernel is initialized at some small, pre-computed value then
  iteratively increased until there are "enough" variants contributing. This
  keeps the prior for one variant from being too strongly influenced by
  extremely close neighbors in annotation space.
}
