% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_bayesian_mpra.R
\name{generate_dist_mat}
\alias{generate_dist_mat}
\title{Generate a distance matrix from a matrix of predictors}
\usage{
generate_dist_mat(predictors, log_distance = TRUE)
}
\arguments{
\item{predictors}{an n x d data frame of predictors}

\item{log_distance}{a logical indicating to use the log1p of the distances (TRUE) or the raw euclidean distances (FALSE)}
}
\description{
Given an nxd matrix of variant annotations, produce an nxn distance matrix
describing the inter-variant distances in annotation space
}
\details{
the predictors need to be ordered by snp_id in the same order as they appear in mpra_data
}
