% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_bayesian_mpra.R
\name{generateDistMat}
\alias{generateDistMat}
\title{Generate a distance matrix from a matrix of predictors}
\usage{
generateDistMat(predictors, log_distance = TRUE)
}
\description{
Generate a distance matrix from a matrix of predictors
}
