% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_bayesian_mpra.R
\name{get_snp_TS_samples}
\alias{get_snp_TS_samples}
\title{Compute transcriptional shift samples}
\usage{
get_snp_TS_samples(sampler_res)
}
\description{
Compute transcriptional shift samples from sampler output
}
