% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_bayesian_mpra.R
\name{quant_norm_parameter}
\alias{quant_norm_parameter}
\title{quantile normalize parameters}
\usage{
quant_norm_parameter(param_name, sample_df)
}
\description{
quantile normalize parameters
}
