% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_bayesian_mpra.R
\name{safelyFitNegBin}
\alias{safelyFitNegBin}
\title{Safely fit a negative binomial}
\usage{
safelyFitNegBin(count_vec)
}
\arguments{
\item{count_vec}{a vector of counts}
}
\value{
a 2-element list of a result or error (the other is NULL)
}
\description{
Safely fit a negative binomial
}
