% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informative_bayesian_mpra.R
\name{bayesian_mpra_analyze}
\alias{bayesian_mpra_analyze}
\title{Bayesian analysis of MPRA data}
\usage{
bayesian_mpra_analyze(mpra_data, predictors, marginal_prior = FALSE)
}
\arguments{
\item{mpra_data}{a data frame of mpra data}

\item{predictors}{a matching data frame of annotations}

\item{marginal_prior}{logical indicating whether or not to disregard the functional predictors and use a marginal prior estimated from the entire assay}
}
\description{
Given MPRA data and a set of predictors, perform a Bayesian analysis of variants using an empirical prior
}
\details{
\code{mpra_data} must meet the following format conditions
  \enumerate { 
    \item one row per barcode 
    \item one column of variant IDs (e.g. rs IDs) 
    \item one column of alleles. These must be character strings of either "ref" or "mut" 
    \item one additional column for every sequencing sample 
    \item column names of plasmid library samples must contain "DNA" (e.g. "DNA_1", "DNA_2", ...) 
    \item column names of samples from transcription products must contain "RNA" (e.g. "RNA_1", "RNA_2", ...) }
}
